% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burglr.R
\name{burglr}
\alias{burglr}
\title{burglg - Stealing code from the web}
\usage{
burglr(urls)
}
\arguments{
\item{urls}{a character vector with web urls indicating the location of an r-script.}
}
\value{
Visual Feedback if the sourcing was successful
}
\description{
With \code{\link{burglr}} you can source r-scripts from the web.
}
\details{
Well, it is as easy as it sounds. The function takes web URL(s) as input
 and evaluates the input in your current R session.
}
\note{
Version 1.0 - f871288d8eae25bcd1dd2bf506f5c47a7cbbc6e4
}
\examples{
burglr(urls = "https://github.com/andrebleier/dive/master/dive.R")

}
\seealso{
For more information and updates see
\href{https://github.com/andrebleier/burglr}{here}.
}
\author{
Andre Bleier
}
