% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_replace.R
\name{char_replace}
\alias{char_replace}
\title{Replacement of non-standard characters}
\usage{
char_replace(
  x,
  to_lower = FALSE,
  trim = TRUE,
  rm_space = FALSE,
  rm_dash = FALSE,
  to_underscore = FALSE
)
}
\arguments{
\item{x}{character vector, contains special characters needing to be replaced.}

\item{to_lower}{logical, forces all characters to lower-case}

\item{trim}{logical, if whitespaces are trimmed from the start and end.
Is ignored with \code{rm_space == TRUE}.}

\item{rm_space}{logical, removes all whitespace characters within 
the string.}

\item{rm_dash}{logical, removes all dash-characters within the string.}

\item{to_underscore}{logical, replaces all whitespace and dash-characters
with underscores.}
}
\value{
a character vector with all non-standard characters replaced by their
standard counterparts
}
\description{
This function replaces non-standard characters (such as the
German Umlaut 'ä') with their standard equivalents (in this case 'ae').
Arguments enable the user to force all characters to lower-case, trim leading 
and trailing whitespace characters or replace all whitespace characters 
and dashes with underscores or remove all whitespace characters within the string.
}
\examples{
x <- "Élizàldë-González Strasse"
char_replace(x)
char_replace
char_replace(x, to_lower = TRUE, to_underscore = TRUE)
char_replace(x, to_lower = TRUE, rm_space = TRUE, rm_dash = TRUE)
}
