% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gc.R
\name{clean_gc}
\alias{clean_gc}
\title{Using gc multiple times}
\usage{
clean_gc(num.gc = 100, threshold = 0.01, verbose = FALSE)
}
\arguments{
\item{num.gc}{a numeric that indicates the maximum number of iterations.}

\item{threshold}{a numeric with the percentage difference. If the change 
in memory size is lower than this, the function stops.}

\item{verbose}{a boolean. If \code{TRUE} information about the run are 
printed.}
}
\description{
Function that cleans the memory by using gc() numerous times.
}
\details{
The function calls \code{gc()} a number of times till the difference of 
the memory size is below the \code{threshold}.
}
\examples{
clean_gc(verbose = TRUE)

}
