% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenstrings.R
\name{evenstrings}
\alias{evenstrings}
\title{splits a string into multiple string with a given maximum length}
\usage{
evenstrings(x = c(), split = " ", char = 80, newlines = FALSE)
}
\arguments{
\item{x}{a string}

\item{split}{a pattern with the splitting symbol. The default is " ".}

\item{char}{a numeric with the maximal length of the result.}

\item{newlines}{a boolean that indicates if the output is seperated by a
newline operator or split into multiple parts.}
}
\value{
returns a vector with the shortend and cutted input string
}
\description{
The function shortens and cuts a given vector to a predefined length.
}
\details{
The splitting is done with \code{\link{strsplit}} with
  \code{type = "after"}.
}
\examples{
x <- "Hello world, this is a test sequence."
evenstrings(x, split = ",", char = 30, newlines = FALSE)
# [1] "Hello world,"              " this is a test sequence."

}
\author{
Jakob Gepp

Jakob Gepp
}
