% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{Finding Pattern in R Files}
\usage{
get_files(dir, pattern = "")
}
\arguments{
\item{dir}{A character string that defines the root directory within which
all R files will be analysed}

\item{pattern}{A character string containing a regular expression to be
matched in all R files found in the specified directory}
}
\value{
Returns a character vector containing all files that match the
  pattern
}
\description{
\code{get_files} searches for matches to argument \code{pattern}
  within each R file that is a descendant of parent directory \code{dir}. It
  works like a global find function that returns all R files within
  \code{dir} that match \code{pattern}. Be careful not to choose a parent
  directory has too many subdirectories and files.
}
\examples{
\dontrun{
# Find all files in . that load packages through library
get_files(dir = ".", pattern = "library")
}
}
\author{
Tobias Krabel
}
