% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequence.R
\name{get_sequence}
\alias{get_sequence}
\title{get start and end indices of sequnces of patterns}
\usage{
get_sequence(x, pattern, minsize = 2L)
}
\arguments{
\item{x}{a vector}

\item{pattern}{the pattern to look for}

\item{minsize}{the minimum length of the repeating pattern}
}
\value{
Returns a matrix with the range of the sequence. Each row representes a
  sequence.
}
\description{
Given a vector \code{x} and a \code{pattern}, the functions returns the
  start and end indices of the sequences with at least \code{minsize}
  repetitions of the \code{pattern}.
}
\examples{
get_sequence(x = c(0,1,1,0,0,0,0,3,0,0,1,0,423,0,0,0,0,0,1,0),
             pattern = 0,
             minsize = 4)

#      min max
# [1,]   4   7
# [2,]  14  18
 
}
\author{
Jakob Gepp
}
