% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{'not in' operator}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are
supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors
are not supported.}
}
\value{
a boolean vector
}
\description{
This negates the '%in%' operator.
}
\examples{
c(1,2,3,4) \%nin\% c(1,2,5)
# [1] FALSE FALSE  TRUE  TRUE

}
\seealso{
\link[base]{match}
}
\author{
Jakob Gepp
}
