% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helfRlein.R
\docType{package}
\name{helfRlein-package}
\alias{helfRlein-package}
\title{This package contains some R helper functions.}
\description{
There is no other purpose to the whole package but to collect and distribute
some of our helper functions which are used over and over again.
}
\author{
Jakob Gepp <jakob.gepp@statworx.com>
}
