% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Combine multiple ggplots}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{multiple ggplots.}

\item{plotlist}{a list with ggplots.}

\item{cols}{numeric. Number of columns in the output plot.}

\item{layout}{a matrix with the layout of the plots. If present, \code{cols}
is ignored.}
}
\description{
This function allows to combine different ggplots into one plot.
}
\details{
ggplot objects can be passed in \code{...}, or to \code{plotlist}
  (as a list of ggplot objects)

If the layout is something like \code{matrix(c(1,2,3,3), nrow = 2, byrow = TRUE)},
then plot 1 will go in the upper left, 2 will go in the upper right, and 3 will go
all the way across the bottom.
}
\seealso{
This is copied from the
  \href{http://www.cookbook-r.com/Graphs/}{Cookbook for R}
}
\author{
Cookbook for R
}
