% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.omit.list.R
\name{na.omit.list}
\alias{na.omit.list}
\title{remove NA from list}
\usage{
\method{na.omit}{list}(y = list(), recursive = FALSE)
}
\arguments{
\item{y}{a list}
}
\value{
returns the list without NA's
}
\details{
This functions removes NA's from a list. With \code{recursive == TRUE} 
   NA's within each listelements are removed as well.
}
\examples{
y <- list(c(1:3), letters[1:4], NA, c(1, NA), list(c(5:6, NA), NA, "A"))
na.omit.list(y, recursive = TRUE)

# [[1]]
# [1] 1 2 3
# 
# [[2]]
# [1] "a" "b" "c" "d"
#
# [[3]]
# [1] 1
#
# [[4]]
# [[4]][[1]]
# [1] 5 6 
#
# [[4]][[2]]
# [1] "A"
}
