% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_omitlist.R
\name{na_omitlist}
\alias{na_omitlist}
\title{remove NA from list}
\usage{
na_omitlist(y = list(), recursive = FALSE)
}
\arguments{
\item{y}{a list}

\item{recursive}{logical. If \code{TRUE}, \code{NA}s within the list's
elements aree removed too.}
}
\value{
Returns the list without \code{NA}s.
}
\description{
This functions removes \code{NA}s from a list. With \code{recursive ==
   TRUE} \code{NA}s within each list's elements are removed as well.
}
\examples{
y <- list(c(1:3), letters[1:4], NA, c(1, NA), list(c(5:6, NA), NA, "A"))
na_omitlist(y, recursive = TRUE)

# [[1]]
# [1] 1 2 3
#
# [[2]]
# [1] "a" "b" "c" "d"
#
# [[3]]
# [1] 1
#
# [[4]]
# [[4]][[1]]
# [1] 5 6
#
# [[4]][[2]]
# [1] "A"

}
\author{
Jakob Gepp
}
