% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_size_in_env.R
\name{object_size_in_env}
\alias{object_size_in_env}
\title{Get size of objects in environment}
\usage{
object_size_in_env(env = .GlobalEnv, unit = "Mb")
}
\arguments{
\item{env}{an environment, the default is the \code{.GlobalEnv}.}

\item{unit}{the units to be used in formatting and printing the size:
"b", "Kb", "Mb", "Gb", "Tb", "Pb", "B", "KB", "MB", "GB", "TB" and "PB".}
}
\value{
Returns a \code{data.table} with the size of objects in the given
environment
}
\description{
Creates a sorted \code{data.table} with the size of objects in a specified
environment.
}
\examples{
DT <- object_size_in_env(env = environment())

}
\author{
Jakob Gepp
}
