% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resave.unique.R
\name{resave_unique}
\alias{resave_unique}
\title{Save only unique data}
\usage{
resave_unique(dt, name = NA, unique.test = TRUE, unique.fac = NULL,
  path = ".")
}
\arguments{
\item{dt}{a data.table}

\item{name}{}

\item{unique.test}{}

\item{unique.fac}{}

\item{path}{a path where the data is saved.}
}
\description{
Function to regoranize large unorganized filse into multiple new
files based on categorical variable.

Function in its current state is generated for MS specific use,
but could be generalized.
}
\examples{


}
