% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.sci.R, R/sci_palette.R
\name{plot.sci}
\alias{plot.sci}
\alias{sci_palette}
\title{Palette for STATWORX Corporate Identity}
\usage{
\method{plot}{sci}(x, ...)

sci_palette(scheme = "old")
}
\arguments{
\item{x}{an object of class \code{sci}.}

\item{...}{arguments to be passed to method}

\item{scheme}{a string that indicates if the "new" or the "old" STATWORX CI is
used. The default is still "old" to not break existing code.}
}
\description{
Palette for STATWORX Corporate Identity colors.
}
\details{
The \code{plot} function gives an example of the colors.
}
\examples{
# Plotting the true underlying effects
plot(sci_palette())

}
\author{
Martin Albers, Jakob Gepp
}
