% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_new_chapter.R
\name{set_new_chapter}
\alias{set_new_chapter}
\title{Insert dashes from current cursor position up to 80 characters}
\usage{
set_new_chapter()
}
\value{
Hyphens inside RStudio
}
\description{
This function inserts dashes from the current cursor position up
  to 80 characters. It can be used for chapter headings for easier
  structuring of your code. You can also assign a shortcut to this function
  in RStudio: Go to "Tools" --> "Modify Keyboard Shortcuts..." and search for
  "Insert Hyphens (---)".
}
\author{
Matthias Nistler
}
