% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sci_palette.R
\name{show_sci_palette}
\alias{show_sci_palette}
\title{show_sci_palette}
\usage{
show_sci_palette()
}
\value{
Returns a plot
}
\description{
\code{show_sci_palette()} plots a graphic with Statworx´s Corporate Identity colors.
}
\author{
Martin Albers
}
