% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusbar.R
\name{statusbar}
\alias{statusbar}
\title{a statusbar for for loops}
\usage{
statusbar(run, max.run, percent.max = 20L, info = run)
}
\arguments{
\item{run}{the iterator of the for loop or an integer with the current
loop number.}

\item{max.run}{either an integer with the maximum number of loops if run is
also a number or a vector with all possible iterations in the right order.}

\item{percent.max}{an integer that indicates how wide the progress
bar is printed.}

\item{info}{a string with additionaly information to be printed at
the end of the line. The default is \code{run}.}
}
\value{
Has no return value, but prints the progress to the console.
}
\description{
This functions prints a progess of a for loop at the console.
}
\examples{

for (i in 1:20) {
  Sys.sleep(0.1)
  statusbar(run = i, max.run = 200, percent.max = 60L)
}

for (i in letters[1:16]) {
  Sys.sleep(0.1)
  statusbar(run = i, max.run = letters[1:16], percent.max = 60L)
}
}
\author{
Jakob Gepp
}
