% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statworx_palette.R
\name{statworx_palette}
\alias{statworx_palette}
\title{a palette based on STATWORX's CI colors}
\usage{
statworx_palette(
  number = length(basecolors),
  reorder = FALSE,
  basecolors = c(1, 2, 3, 5, 10)
)
}
\arguments{
\item{number}{a numeric with the number of colors.
The default is the length of \code{basecolors}.}

\item{reorder}{a boolean if \code{TRUE} the order will be sampled,
which can make groups next to each other a little bit more separable.}

\item{basecolors}{a numeric vector with the used color indices of 
\code{\link{sci_palette}}.}
}
\description{
A palette based on STATWORX's CI colors.
}
\examples{
number <- 16
plot(data.frame(x =  1:number, y = 1), 
     col = statworx_palette(number, FALSE),
     pch = 16, cex = 5)

}
\author{
Jakob Gepp
}
