% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit.R
\name{strsplit}
\alias{strsplit}
\title{Improved strsplit function}
\usage{
strsplit(x, split, type = "remove", perl = FALSE, ...)
}
\arguments{
\item{x}{character vector, each element of which is to be split. Other
inputs, including a factor, will give an error.}

\item{split}{character vector (or object which can be coerced to such)
containing regular expression(s) (unless \code{fixed = TRUE}) to use for
splitting. If empty matches occur, in particular if \code{split} has length
0, \code{x} is split into single characters. If \code{split} has length
greater than 1, it is recycled along \code{x}.}

\item{type}{a charachter. Either to \code{"remove"} or keep the delimiter
\code{"before"}, \code{"between"} or \code{"remove"} the split.}

\item{perl}{logical. Should Perl-compatible regexps be used? Is \code{TRUE}
for all but \code{"remove"}.}

\item{...}{other inputs passed along to \code{\link[base]{strsplit}}}
}
\value{
A list of the same length as \code{x}, the i-th element of which contains
  the vector of splits of \code{x[i]}.
}
\description{
This functions uses \code{\link[base]{strsplit}} and adds the
possibility to split and keep the delimiter after or before the given split.
Or you can split between two given splits and keep both.
}
\examples{
x <- c("3D/MON&SUN")
strsplit(x, "[/&]")
# [[1]]
# [1] "3D"  "MON" "SUN"

strsplit(x, "[/&]", type = "before")
# [[1]]
# [1] "3D"  "/MON" "&SUN"

strsplit(x, "[/&]", type = "after")
# [[1]]
# [1] "3D/"  "MON&" "SUN"

x <- c("3D/MON&SUN 2D/MON&SUN")
strsplit(x, split("/", "M"))
# [[1]]
# [1] "3D"         "MON&SUN 2D" "MON&SUN"

}
\seealso{
\code{\link[base]{strsplit}} or
  \href{https://stackoverflow.com/questions/15575221}{stackoverflow}
   for more details.
}
\author{
Jakob Gepp
}
