% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_network}
\alias{get_network}
\title{flowchart of R projects}
\usage{
get_network(
  dir = NULL,
  variations = c("<- function", " <- function", "<-function", " <-function"),
  pattern = "\\\\.R$",
  simplify = FALSE,
  all_scripts = NULL,
  use_internals = TRUE,
  exclude = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dir}{a path that includes the functions}

\item{variations}{a character vector with the function's definition string.
The default is c(" <- function", "<- function", "<-function").}

\item{pattern}{a string with the file suffix - default is "\\.R$".}

\item{simplify}{a boolean, if \code{TRUE} function with no connections are
removed from the plot.}

\item{all_scripts}{a named list with script. The given path is dominant. This
is mainly used for debugging purposes.}

\item{use_internals}{a boolean, if \code{FALSE} than only functions
that are also a file's name are used. This should respond to internal
functions. The default is \code{TRUE}.}

\item{exclude}{a vector with folder's or function's names, that are excluded
from the network creation. This is done by a regex, so it will remove
everything that contains these words.}

\item{verbose}{a boolean setting the debugging prints.}
}
\value{
Returns an object with the adjacency matrix \code{$matrix},
  an igraph object \code{$igraph}, a table for the edges \code{$edge_dt},
  a table for the nodes \code{$node_dt} and
  a networkD3 plot \code{$networkD3}.
}
\description{
With this function a network plot of the connections of the functions
in a given path can be created.
}
\note{
TODO: list with exclude files and comments ' ' in one line

TODO: maybe return plot
}
\examples{
\dontrun{
net <- get_network(dir = "R/", simplify = TRUE)
g1 <- net$igraph
plot(g1)
}

}
\seealso{
For more information check out
\href{https://www.statworx.com/de/blog/flowcharts-of-functions/}{our blog}.
Also there is a README on how to use this function
 \href{https://github.com/STATWORX/blog/tree/master/flowchart}{here}.
}
\author{
Jakob Gepp
}
