% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnetwork.R
\name{getnetwork}
\alias{getnetwork}
\title{flowchart of R projects}
\usage{
getnetwork(dir, variations = c(" <- function", "<- function", "<-function"),
  pattern = "\\\\.R")
}
\arguments{
\item{dir}{a path that includes the functions}

\item{variations}{a charachter vector with the functions definition string.
The default is c(" <- function", "<- function", "<-function").}

\item{pattern}{a string with the file suffix - default is "\\.R".}
}
\value{
Returns an object with the adjacency matrix \code{$matrix} and 
  and igraph object \code{$igraph}.
}
\examples{

}
