% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helfRlein-package.R
\docType{package}
\name{helfRlein-package}
\alias{helfRlein}
\alias{helfRlein-package}
\title{helfRlein: R Helper Functions}
\description{
This package contains a collection of R helper functions which
  we repeatedly use at \href{https://www.statworx.com/}{STATWORX} to ease our
  programming life and our daily work and projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/STATWORX/helfRlein}
  \item Report bugs at \url{https://github.com/STATWORX/helfRlein/issues}
}

}
\author{
\strong{Maintainer}: Jakob Gepp \email{jakob.gepp@statworx.com}

Authors:
\itemize{
  \item Daniel Luettgau \email{daniel.luettgau@statworx.com}
  \item Andre Bleier \email{andre.bleier@statworx.com}
  \item Tobias Krabel \email{tobias.krabel@statworx.com}
  \item Martin Albers \email{martin.albers@statworx.com}
  \item Lukas Feick \email{lukas.feick@statworx.com}
}

Other contributors:
\itemize{
  \item Markus Berroth \email{markus.berroth@statworx.com} [contributor]
  \item Robin Kraft \email{robin.kraft@statworx.com} [contributor]
  \item David Schlepps \email{david.schlepps@statworx.com} [contributor]
  \item Oliver Guggenbuehl \email{oliver.guggenbuehl@statworx.com} [contributor]
}

}
\keyword{internal}
