% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect2.R
\name{intersect2}
\alias{intersect2}
\title{intersect for multiple input vectors}
\usage{
intersect2(...)
}
\arguments{
\item{...}{vectors to check for intersect}
}
\value{
Returns the intersect of all given input vectors.
}
\description{
Function to check the intersect within multiple vectors.
}
\examples{
intersect2(c(1:3), c(1:4), c(1:2))
# [1] 1 2

}
