% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{combine multiple ggplots}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{multiple ggplots.}

\item{plotlist}{a list with ggplots.}

\item{cols}{numeric. Number of columns in the output plot.}

\item{layout}{a matrix with the layout of the plots.}
}
\description{
Functions that allows to combine diffrent ggplots into one plot.
}
\details{
ggplot objects can be passed in ..., or to plotlist (as a list of
ggplot objects) - cols: Number of columns in layout - layout: A matrix
specifying the layout. If present, 'cols' is ignored.

If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE), then
plot 1 will go in the upper left, 2 will go in the upper right, and 3 will go
all the way across the bottom.
}
\seealso{
This is copied from the
 \href{http://www.cookbook-r.com/Graphs/
 Multiple_graphs_on_one_page_(ggplot2)/}{Cookbook for R}
}
\author{
Cookbook for R
}
