% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fs.R
\name{print_fs}
\alias{print_fs}
\title{print file structure}
\usage{
print_fs(path = ".", silent = FALSE)
}
\arguments{
\item{path}{a folder path.}

\item{silent}{a logical value. If \code{TRUE} the return value is a data.tree
with the file structure and nothing is printed.}
}
\value{
Either the file structure gets printed or returned.
}
\description{
Prints the directories and files of the given path.
}
\examples{

print_fs(path = ".")

}
\author{
Jakob Gepp
}
