% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read multiple files into one \code{data.table}}
\usage{
read_files(files, fun = readLines, ...)
}
\arguments{
\item{files}{paths to files to be read in}

\item{fun}{function to load files in}

\item{...}{additional arguments to be passed to \code{fun}}
}
\value{
combined \code{data.table} of the input files
}
\description{
This function reads multiple files into one \code{data.table}
}
\examples{

\dontrun{
# this needs files
read_files(files, fun = readRDS)
read_files(files, fun = readLines)
read_files(files, fun = read.csv, sep = ";")
}

}
\author{
Jakob Gepp
}
