% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_palette.R
\name{sci_palette}
\alias{sci_palette}
\title{Palette for STATWORX Corporate Identity}
\usage{
sci_palette(number = NULL)
}
\arguments{
\item{number}{a numeric \code{(1-9)} with the number of colors}
}
\description{
Palette for Statworx's Corporate Identity colors.
}
\details{
The \code{plot} function gives an example of the colors.
}
\examples{
sci_palette(3) # plotting three colors

}
\author{
Martin Albers, Jakob Gepp
}
