% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.sci.R, R/sci_palette.R
\name{plot.sci}
\alias{plot.sci}
\alias{sci_palette}
\title{Palette for STATWORX Corporate Identity}
\usage{
\method{plot}{sci}(x, ...)

sci_palette(number = NULL)
}
\arguments{
\item{x}{an object of class \code{sci}.}

\item{...}{arguments to be passed to method}

\item{number}{a numeric \code{(1-9)} with the number of colors}
}
\description{
Palette for Statworx's Corporate Identity colors.
}
\details{
The \code{plot} function gives an example of the colors.
}
\examples{
# Plotting the true underlying effects
plot(sci_palette())

sci_palette(3) # plotting three colors

}
\author{
Martin Albers, Jakob Gepp
}
