% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_palette.R
\name{sci_palette}
\alias{sci_palette}
\title{Palette for STATWORX Corporate Identity}
\usage{
sci_palette()
}
\description{
Palette for Statworx´s Corporate Identity colors.
}
\examples{
plot(c(1,2,3), c(1,2,3), col = sci_palette()[3], pch = 16)

}
\author{
Martin Albers
}
