% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_new_chapter.R
\name{set_new_chapter}
\alias{set_new_chapter}
\title{Insert dashes from courser position up to 80 characters}
\usage{
set_new_chapter()
}
\value{
Hyphens inside RStudio
}
\description{
Insert dashes from courser position up to 80 characters
}
\examples{
## NOT RUN
}
\author{
Matthias Nistler
}
