% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusbar.R
\name{statusbar}
\alias{statusbar}
\title{A statusbar for for loops}
\usage{
statusbar(run, max.run, width = 20L, info = run, percent.max = width)
}
\arguments{
\item{run}{the iterator of the \code{for} loop or an integer with the current
loop number.}

\item{max.run}{either an integer with the maximum number of loops if \code{run} is
also a number, or a vector with all possible iterations in the correct order.}

\item{width}{an integer that indicates how wide the progress
bar is printed.}

\item{info}{a string with additional information to be printed at
the end of the line. The default is \code{run}.}

\item{percent.max}{is deprecated, use width instead.}
}
\value{
Has no return value, but prints the progress to the console.
}
\description{
This functions prints the progess of a \code{for} loop to the console.
}
\examples{

for (i in 1:20) {
  Sys.sleep(0.1)
  statusbar(run = i, max.run = 200, width = 60L)
}

for (i in letters[1:16]) {
  Sys.sleep(0.1)
  statusbar(run = i, max.run = letters[1:16], width = 60L)
}
}
\author{
Jakob Gepp
}
