% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statworx_palette.R
\name{statworx_palette}
\alias{statworx_palette}
\title{a palette based on STATWORX's CI colors}
\usage{
statworx_palette(number = 4, reorder = FALSE)
}
\arguments{
\item{number}{a numeric \code{(1-9)} with the number of colors}

\item{reorder}{a boolean if \code{TRUE} the order will be sampled,
which can make groups next to each other a little bit more separable.}
}
\description{
A palette based on STATWORX's CI colors.
}
\examples{
number <- 16
plot(data.frame(x =  1:number, y = 1), 
     col = statworx_palette(number, TRUE),
     pch = 16, cex = 5)

}
\author{
Jakob Gepp
}
