% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_na.R
\name{to_na}
\alias{to_na}
\title{replace NaN and Inf with NA}
\usage{
to_na(x)
}
\arguments{
\item{x}{vector}
}
\value{
Returns vector with with replaced \code{NA}values.
}
\description{
Takes out \code{NaN} and \code{Inf} and replaces them with \code{NA}
}
\note{
--- Idea for improvement

Add args to flexible select which scenarios should be set NA

  - nan, infinite, other defined values
}
\examples{
test <- list(a = c("a", "b", NA),
             b = c(NaN, 1,2, -Inf),
             c = c(TRUE, FALSE, NA))

lapply(test, to_na)

## Output
# $a
# [1] "a" "b" NA

# $b
# [1] NA  1  2 NA

# $c
# [1] TRUE FALSE NA

}
\author{
Daniel Luettgau

Daniel Luettgau
}
