% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim leading or trailing whitespaces}
\usage{
trim(x, lead = TRUE, trail = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{lead}{a boolean. If \code{TRUE} leading whitespaces will be trimed}

\item{trail}{a boolean. If \code{TRUE} trailing whitespaces will be trimed}
}
\value{
a character vector
}
\description{
Trim leading or trailing whitespaces
}
\examples{
x <- c("  Hello world!", "  Hello world! ", "Hello world! ")
trim(x)
# [1] "Hello world!" "Hello world!" "Hello world!"

}
\seealso{
Since R 3.5.1 there is \code{\link[base]{trimws}} in the base
  package.
}
\author{
Jakob Gepp
}
